refactor SceneBuilder
